* Patched sourcecode with modifications made in OpenTibia subversion branch: avesta.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@438 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ npc.cpp
@@ -1182,7 +1182,7 @@ void Npc::onThink(uint32_t interval)
 			if(npcState->prevInteraction == 0)
 				npcState->prevInteraction = OTSYS_TIME();
 
-			if(idleTime > 0 && (OTSYS_TIME() - npcState->prevInteraction) > (uint64_t)(idleTime * 1000))
+			if(idleTime > 0 && (OTSYS_TIME() - npcState->prevInteraction) > idleTime * 1000)
 			{
 				idleTimeout = true;
 				closeConversation = true;
